fix: удалить статическое внедрение клиента статистики
