--- stat-service/client/src/main/java/ru/practicum/ClientFactory.java
@@ -1,7 +1,0 @@
-package ru.practicum;
-
-public class ClientFactory {
-    public static StatClient getDefaultClient(String url) {
-        return new StatClientImpl(url);
-    }
-}
